refactor: map 관련 상수 파일 분리
