+++ src/components/Map/index.tsx
@@ -19,17 +19,21 @@ import DustLevel from '@/components/common/DustLevel';
 import DustState from '@/components/common/DustState';
 import { getAllLocation } from '@/apis/location';
 import { getSidoDustInfos, getCityDustInfos } from '@/apis/dustInfo';
-import { getDustScaleColor } from '@/utils/map';
-import { FINE_DUST, ULTRA_FINE_DUST } from '@/utils/constants';
 import {
-  COLOR_MARKER_MOUSE_OUT,
+  FINE_DUST,
+  ULTRA_FINE_DUST,
+  DUST_GRADE,
+  DUST_SCALE_COLOR,
+  CITY_ZOOM_LEVEL,
+  MAX_ZOOM_LEVEL,
   COLOR_MARKER_MOUSE_OVER,
-  ZINDEX_MARKER_MOUSE_OUT,
+  COLOR_MARKER_MOUSE_OUT,
   ZINDEX_MARKER_MOUSE_OVER,
-} from '@/utils/map';
+  ZINDEX_MARKER_MOUSE_OUT,
+} from '@/utils/constants';
 import useMap from '@/hooks/useMap';
-import { MAX_ZOOM_LEVEL, CITY_ZOOM_LEVEL } from '@/utils/map';
 import type { CityDustInfo } from '@/types/dust';
+import { getDustAverageGrade } from '@/utils/dustGrade';
 
 declare global {
   interface Window {
@@ -80,13 +84,24 @@ const Map = () => {
   useEffect(() => {
     if (!map || !sidoDustInfos || !allLocation) return;
 
-    sidoDustInfos.forEach(({ sidoName, fineDustScale, ultraFineDustScale }) => {
-      const { latitude, longitude } = allLocation.filter(
-        (scale) => scale.sidoName === sidoName
-      )[0];
+    sidoDustInfos.forEach(
+      ({
+        sidoName,
+        fineDustScale,
+        fineDustGrade,
+        ultraFineDustScale,
+        ultraFineDustGrade,
+      }) => {
+        const { latitude, longitude } = allLocation.filter(
+          (scale) => scale.sidoName === sidoName
+        )[0];
 
-      const backgroundColor = getDustScaleColor(fineDustScale);
-      const template = `
+        const averageGrade = getDustAverageGrade(
+          fineDustGrade,
+          ultraFineDustGrade
+        );
+        const backgroundColor = DUST_SCALE_COLOR[DUST_GRADE[averageGrade]];
+        const template = `
           <div class="dust-info-marker" style="background-color: ${backgroundColor};">
             <p class="city-name">${sidoName}</p>
             <div class="dust-info">
@@ -96,14 +111,15 @@ const Map = () => {
             </div>
           </div>`;
 
-      const marker = new kakao.maps.CustomOverlay({
-        clickable: true,
-        position: new kakao.maps.LatLng(latitude, longitude),
-        content: template,
-      });
+        const marker = new kakao.maps.CustomOverlay({
+          clickable: true,
+          position: new kakao.maps.LatLng(latitude, longitude),
+          content: template,
+        });
 
-      sidoDustInfoMarkers.push(marker);
-    });
+        sidoDustInfoMarkers.push(marker);
+      }
+    );
 
     sidoDustInfoMarkers.forEach((marker) => {
       marker.setMap(map);
@@ -133,15 +149,20 @@ const Map = () => {
       ({
         cityName,
         fineDustScale,
-        ultraFineDustScale,
         fineDustGrade,
+        ultraFineDustScale,
         ultraFineDustGrade,
       }) => {
         geocoder.addressSearch(cityName, (result, status) => {
           if (status === kakao.maps.services.Status.OK) {
             const latitude = Number(result[0].y);
             const longitude = Number(result[0].x);
-            const backgroundColor = getDustScaleColor(fineDustScale);
+
+            const averageGrade = getDustAverageGrade(
+              fineDustGrade,
+              ultraFineDustGrade
+            );
+            const backgroundColor = DUST_SCALE_COLOR[DUST_GRADE[averageGrade]];
             const template = `
                   <div class="dust-info-marker" id="${cityName}" data-finedustgrade="${fineDustGrade}" data-ultrafinedustgrade="${ultraFineDustGrade}" style="background-color: ${backgroundColor};" >
                     <span>${fineDustScale}/${ultraFineDustScale}</span>
