+++ src/utils/constants/index.ts
@@ -2,3 +2,13 @@ export { SIDO_GROUP } from '@/utils/constants/city';
 export { INIT_LOCATION, CENTER_LOCATION } from '@/utils/constants/location';
 export { FINE_DUST, ULTRA_FINE_DUST, DUST_GRADE } from '@/utils/constants/dust';
 export { ROUTE } from '@/utils/constants/route';
+export {
+  INIT_ZOOM_LEVEL,
+  CITY_ZOOM_LEVEL,
+  MAX_ZOOM_LEVEL,
+  COLOR_MARKER_MOUSE_OVER,
+  COLOR_MARKER_MOUSE_OUT,
+  ZINDEX_MARKER_MOUSE_OVER,
+  ZINDEX_MARKER_MOUSE_OUT,
+  DUST_SCALE_COLOR,
+} from '@/utils/constants/map';
