+++ src/utils/constants/map.ts
@@ -1,0 +1,15 @@
+export const INIT_ZOOM_LEVEL = 5;
+export const CITY_ZOOM_LEVEL = 9;
+export const MAX_ZOOM_LEVEL = 12;
+
+export const COLOR_MARKER_MOUSE_OVER = 'yellow';
+export const COLOR_MARKER_MOUSE_OUT = 'white';
+export const ZINDEX_MARKER_MOUSE_OVER = '100';
+export const ZINDEX_MARKER_MOUSE_OUT = '0';
+
+export const DUST_SCALE_COLOR = {
+  DANGER: '#e64746',
+  BAD: '#fda60d',
+  NORMAL: '#03c73c',
+  GOOD: '#30a2ff',
+};
